Feat: 직관 일기 관련 기능

+++ src/main/java/com/boot/gugi/repository/StadiumRepository.java
@@ -3,6 +3,6 @@ package com.boot.gugi.repository;
 import com.boot.gugi.model.Stadium;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StadiumRepository extends JpaRepository<Stadium, Long> {
+public interface StadiumRepository extends JpaRepository<Stadium, Integer> {
     Stadium findByStadiumCode(Integer stadiumCode);
 }
